refactor: [ZN-411] Adjust auth API actions collection
