+++ src/api/actions/auth/auth.mutations.ts
@@ -1,9 +1,14 @@
 import { AxiosInstance } from 'axios';
 
-import { LoginMutationArguments, LoginMutationResponse } from './authActions.types';
+import {
+  LoginMutationArguments,
+  LoginMutationResponse,
+  // MUTATION_TYPE_IMPORTS
+} from './auth.types';
 
 export const authMutations = {
   loginMutation: (client: AxiosInstance) => async (body: LoginMutationArguments) => {
     return (await client.post<LoginMutationResponse>('/authorize', body)).data;
   },
+  // MUTATION_FUNCTIONS_SETUP
 };
