+++ src/api/actions/auth/auth.queries.ts
@@ -1,7 +1,13 @@
-import { stringify } from 'qs';
 import { AxiosInstance } from 'axios';
+import { stringify } from 'qs';
 
-import { GetMeQueryResponse, GetUsersInfiniteArgs, GetUsersListArgs, GetUsersResponse } from './authActions.types';
+import {
+  GetMeQueryResponse,
+  GetUsersInfiniteArgs,
+  GetUsersListArgs,
+  GetUsersResponse,
+  // QUERY_TYPE_IMPORTS
+} from './auth.types';
 
 export const authQueries = {
   getCurrentUser: (client: AxiosInstance) => async () => {
@@ -19,4 +25,5 @@ export const authQueries = {
       const queryParams = stringify({ page, count: 5 }, { addQueryPrefix: true });
       return (await client.get<GetUsersResponse>(`/users/${queryParams}`)).data;
     },
+  // QUERY_FUNCTIONS_SETUP
 };
