+++ src/api/actions/index.ts
@@ -1,14 +1,17 @@
-import { authMutations } from './auth/authActions.mutations';
-import { authQueries } from './auth/authActions.queries';
+import { authMutations } from './auth/auth.mutations';
+import { authQueries } from './auth/auth.queries';
+// API_COLLECTION_IMPORTS
 
 export const queries = {
   ...authQueries,
+  // API_COLLECTION_QUERIES
 } as const;
 
 export type AxiosQueriesType = typeof queries;
 
 export const mutations = {
   ...authMutations,
+  // API_COLLECTION_MUTATIONS
 } as const;
 
 export type AxiosMutationsType = typeof mutations;
