+++ src/hooks/useUser/useUser.ts
@@ -1,6 +1,6 @@
 import { UseQueryOptions } from '@tanstack/react-query';
 
-import { GetMeQueryResponse } from '../../api/actions/auth/authActions.types';
+import { GetMeQueryResponse } from 'api/actions/auth/auth.types';
 import { useQuery } from '../useQuery/useQuery';
 
 export const useUser = (options?: UseQueryOptions<GetMeQueryResponse>) => {
